+++ CHANGELOG.md
@@ -33,6 +33,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a custom conrod rich text widget to render in-line images
 - Added 5 new Firestaff abilities.
 - Server-side Prometheus metrics for entity deaths, locations, and causes of death
+- New river ambience
+- Softer rain/ambience when indoors
   
 ### Changed
 
@@ -57,6 +59,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Chat command autocomplete now works correctly again.
 - Resources in spots are now tracked by rtsim.
 - Fix Orientation of Wildskin Drum's Thumbnail
+- Fast river sfx
+- All NPCs don't go to bed & wake up at the same time.
 
 ## [0.18.0] - 2026-01-23